LayoutTests/inspector/runtime/setSavedResultAlias-expected.txt

Tests for the Runtime.setSavedResultAlias command.


== Running test suite: Runtime.setSavedResultAlias
-- Running test case: Runtime.setSavedResultAlias
Saving value...

Getting saved value at '$1'...
PASS: Saved value should match $1.
Getting saved value at 'temp1'...
PASS: Should produce an exception.
Error: ReferenceError: Can't find variable: temp1

Changing saved result alias to "temp"...

Getting saved value at '$1'...
PASS: Saved value should still match $1.
Getting saved value at 'temp1'...
PASS: Saved value should now match temp1.

Changing saved result alias to ""...

Getting saved value at '$1'...
PASS: Saved value should still match $1.
Getting saved value at 'temp1'...
PASS: Should produce an exception.
Error: ReferenceError: Can't find variable: temp1
